requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
groq>=0.4.0
pytest>=8.0.0
httpx>=0.27.0,<0.28.0